+++ src/main/java/bfst20/mapdrawer/osm/OSMNode.java
@@ -10,8 +10,8 @@ public class OSMNode implements LongSupplier, Serializable {
     private static final long serialVersionUID = 1L;
 
     private final long id;
-    private final double lon; // x
-    private final double lat; // y
+    private final float lon; // x
+    private final float lat; // y
     
     private String address;
 
@@ -19,7 +19,7 @@ public class OSMNode implements LongSupplier, Serializable {
 
     private int numberForGraph;
 
-    public OSMNode(long id, double lon, double lat, int numberForGraph, String address) {
+    public OSMNode(long id, float lon, float lat, int numberForGraph, String address) {
         this.id = id;
         this.lon = lon;
         this.lat = lat;
@@ -38,11 +38,11 @@ public class OSMNode implements LongSupplier, Serializable {
         return id;
     }
 
-    public double getLon() {
+    public float getLon() {
         return lon;
     }
 
-    public double getLat() {
+    public float getLat() {
         return lat;
     }
     
@@ -66,10 +66,10 @@ public class OSMNode implements LongSupplier, Serializable {
      * so it can only be used for condition checks (because the relation holds, so ex. distance1 < distance2 still works
      * - but the numbers themselves are wrong). Normal calculation is sqrt(a^2 + b^2), this one is only a^2 + b^2.
      */
-    public double distanceSq(Point2D point) {
+    public float distanceSq(Point2D point) {
         double dX = point.getX() - lon;
         double dY = point.getY() - lat;
-        return dX * dX + dY * dY;
+        return (float) (dX * dX + dY * dY);
     }
     
     public void setRoad(String road) {
